checkin progress
